2.5 final antes de listas

+++ grafo.h
@@ -9,6 +9,7 @@
 #include <set>
 #include <unistd.h>
 #include <string>
+#include <stdint.h>
 
 
 
@@ -16,23 +17,23 @@ class Grafo
 {
 protected:
     std::vector<std::string> verticesArreglo;
-    std::vector<std::vector<short>> matrixAristas;
-    short tamanoPalabras;
+    std::vector<std::vector<uint8_t>> matrixAristas;
+    uint8_t tamanoPalabras;
 public:
     void iniciarMatrix(int tamano);
     void insertarVertice(std::vector<std::string> listaPalabras);
     int buscarVerticePosicion(std::string verticeBuscar);
-    void insetarArista(std::string puntoInicio, std::string puntoFinal, short costo);
-    std::string buscarVerticePorPosicion(short posicion);
+    void insetarArista(std::string puntoInicio, std::string puntoFinal, uint8_t costo);
+    std::string buscarVerticePorPosicion(uint8_t posicion);
     int obtenerTamanoArregloVectores();
     void imprimirVertices();
     void imprimirMatrix();
-    void fijarTamanoPalabras(short tamano);
-    short obtenerTamanoPalabras();
+    void fijarTamanoPalabras(uint8_t tamano);
+    uint8_t obtenerTamanoPalabras();
     //Dijkstra
-    std::vector<short> encontrarVecinosVertice(short vertice);
-    std::vector<std::string> dijkstra(short inicio, short final);
-    short buscarPosicionVerticeMenorCosto(std::vector<short> arregloDistancias, std::vector<short> arregloVerticesFuncion);
+    std::vector<uint8_t> encontrarVecinosVertice(uint8_t vertice);
+    std::vector<std::string> dijkstra(uint8_t inicio, uint8_t final);
+    uint8_t buscarPosicionVerticeMenorCosto(std::vector<uint8_t> arregloDistancias, std::vector<uint8_t> arregloVerticesFuncion);
 };
 
 #include "grafo.hxx"
